+++ packages/flutter_tools/lib/executable.dart
@@ -50,6 +50,8 @@ import 'src/globals.dart' as globals;
 // Files in `isolated` are intentionally excluded from google3 tooling.
 import 'src/isolated/build_targets.dart';
 import 'src/isolated/mustache_template.dart';
+import 'src/isolated/native_assets/native_assets.dart';
+import 'src/isolated/native_assets/test/native_assets.dart';
 import 'src/isolated/resident_web_runner.dart';
 import 'src/pre_run_validator.dart';
 import 'src/project_validator.dart';
@@ -177,6 +179,7 @@ List<FlutterCommand> generateCommands({
     platform: globals.platform,
     processInfo: globals.processInfo,
     fileSystem: globals.fs,
+    nativeAssetsBuilder: const HotRunnerNativeAssetsBuilderImpl(),
   ),
   BuildCommand(
     artifacts: globals.artifacts!,
@@ -237,10 +240,17 @@ List<FlutterCommand> generateCommands({
     platform: globals.platform,
     featureFlags: featureFlags,
   ),
-  RunCommand(verboseHelp: verboseHelp),
+  RunCommand(
+    verboseHelp: verboseHelp,
+    nativeAssetsBuilder: const HotRunnerNativeAssetsBuilderImpl(),
+  ),
   ScreenshotCommand(fs: globals.fs),
   ShellCompletionCommand(),
-  TestCommand(verboseHelp: verboseHelp, verbose: verbose),
+  TestCommand(
+    verboseHelp: verboseHelp,
+    verbose: verbose,
+    nativeAssetsBuilder: const TestCompilerNativeAssetsBuilderImpl(),
+  ),
   UpgradeCommand(verboseHelp: verboseHelp),
   SymbolizeCommand(
     stdio: globals.stdio,
