+++ packages/flutter_tools/lib/src/build_system/targets/native_assets.dart
@@ -7,18 +7,18 @@ import 'package:native_assets_cli/native_assets_cli_internal.dart' show Asset;
 import 'package:package_config/package_config_types.dart';
 
 import '../../android/gradle_utils.dart';
-import '../../android/native_assets.dart';
 import '../../base/common.dart';
 import '../../base/file_system.dart';
 import '../../base/platform.dart';
 import '../../build_info.dart';
 import '../../dart/package_map.dart';
-import '../../ios/native_assets.dart';
-import '../../linux/native_assets.dart';
-import '../../macos/native_assets.dart';
+import '../../isolated/native_assets/android/native_assets.dart';
+import '../../isolated/native_assets/ios/native_assets.dart';
+import '../../isolated/native_assets/linux/native_assets.dart';
+import '../../isolated/native_assets/macos/native_assets.dart';
+import '../../isolated/native_assets/native_assets.dart';
+import '../../isolated/native_assets/windows/native_assets.dart';
 import '../../macos/xcode.dart';
-import '../../native_assets.dart';
-import '../../windows/native_assets.dart';
 import '../build_system.dart';
 import '../depfile.dart';
 import '../exceptions.dart';
