+++ packages/flutter_tools/lib/src/commands/attach.dart
@@ -71,6 +71,7 @@ class AttachCommand extends FlutterCommand {
     required Platform platform,
     required ProcessInfo processInfo,
     required FileSystem fileSystem,
+    HotRunnerNativeAssetsBuilder? nativeAssetsBuilder,
   }) : _hotRunnerFactory = hotRunnerFactory ?? HotRunnerFactory(),
        _stdio = stdio,
        _logger = logger,
@@ -78,7 +79,8 @@ class AttachCommand extends FlutterCommand {
        _signals = signals,
        _platform = platform,
        _processInfo = processInfo,
-       _fileSystem = fileSystem {
+       _fileSystem = fileSystem,
+       _nativeAssetsBuilder = nativeAssetsBuilder {
     addBuildModeFlags(verboseHelp: verboseHelp, defaultToRelease: false, excludeRelease: true);
     usesTargetOption();
     usesPortOptions(verboseHelp: verboseHelp);
@@ -150,6 +152,7 @@ class AttachCommand extends FlutterCommand {
   final Platform _platform;
   final ProcessInfo _processInfo;
   final FileSystem _fileSystem;
+  final HotRunnerNativeAssetsBuilder? _nativeAssetsBuilder;
 
   @override
   final String name = 'attach';
@@ -416,6 +419,7 @@ known, it can be explicitly provided to attach via the command-line, e.g.
           device: device,
           flutterProject: flutterProject,
           usesIpv6: usesIpv6,
+          nativeAssetsBuilder: _nativeAssetsBuilder,
         );
         late AppInstance app;
         try {
@@ -449,6 +453,7 @@ known, it can be explicitly provided to attach via the command-line, e.g.
           device: device,
           flutterProject: flutterProject,
           usesIpv6: usesIpv6,
+          nativeAssetsBuilder: _nativeAssetsBuilder,
         );
         final Completer<void> onAppStart = Completer<void>.sync();
         TerminalHandler? terminalHandler;
@@ -505,6 +510,7 @@ known, it can be explicitly provided to attach via the command-line, e.g.
     required Device device,
     required FlutterProject flutterProject,
     required bool usesIpv6,
+    required HotRunnerNativeAssetsBuilder? nativeAssetsBuilder,
   }) async {
     final BuildInfo buildInfo = await getBuildInfo();
 
@@ -539,6 +545,7 @@ known, it can be explicitly provided to attach via the command-line, e.g.
           ipv6: usesIpv6,
           flutterProject: flutterProject,
           nativeAssetsYamlFile: stringArg(FlutterOptions.kNativeAssetsYamlFile),
+          nativeAssetsBuilder: _nativeAssetsBuilder,
           analytics: analytics,
         )
       : ColdRunner(
@@ -573,6 +580,7 @@ class HotRunnerFactory {
     bool ipv6 = false,
     FlutterProject? flutterProject,
     String? nativeAssetsYamlFile,
+    required HotRunnerNativeAssetsBuilder? nativeAssetsBuilder,
     required Analytics analytics,
   }) => HotRunner(
     devices,
@@ -586,6 +594,7 @@ class HotRunnerFactory {
     stayResident: stayResident,
     ipv6: ipv6,
     nativeAssetsYamlFile: nativeAssetsYamlFile,
+    nativeAssetsBuilder: nativeAssetsBuilder,
     analytics: analytics,
   );
 }
