+++ packages/flutter_tools/lib/src/isolated/native_assets/README.md
@@ -1,0 +1,4 @@
+The directory structure in here mirrors the directory structure in lib/src/.
+
+The native assets feature lives in lib/src/isolated/ because this code is not
+used in other build systems.
