+++ packages/flutter_tools/lib/src/isolated/native_assets/ios/native_assets.dart
@@ -9,9 +9,9 @@ import 'package:native_assets_cli/native_assets_cli_internal.dart'
 import 'package:native_assets_cli/native_assets_cli_internal.dart'
     as native_assets_cli;
 
-import '../base/file_system.dart';
-import '../build_info.dart';
-import '../globals.dart' as globals;
+import '../../../base/file_system.dart';
+import '../../../build_info.dart';
+import '../../../globals.dart' as globals;
 
 import '../macos/native_assets_host.dart';
 import '../native_assets.dart';
