+++ packages/flutter_tools/lib/src/run_hot.dart
@@ -22,7 +22,6 @@ import 'dart/package_map.dart';
 import 'devfs.dart';
 import 'device.dart';
 import 'globals.dart' as globals;
-import 'native_assets.dart';
 import 'project.dart';
 import 'reporting/reporting.dart';
 import 'resident_runner.dart';
@@ -93,13 +92,13 @@ class HotRunner extends ResidentRunner {
     StopwatchFactory stopwatchFactory = const StopwatchFactory(),
     ReloadSourcesHelper reloadSourcesHelper = defaultReloadSourcesHelper,
     ReassembleHelper reassembleHelper = _defaultReassembleHelper,
-    NativeAssetsBuildRunner? buildRunner,
+    HotRunnerNativeAssetsBuilder? nativeAssetsBuilder,
     String? nativeAssetsYamlFile,
     required Analytics analytics,
   })  : _stopwatchFactory = stopwatchFactory,
         _reloadSourcesHelper = reloadSourcesHelper,
         _reassembleHelper = reassembleHelper,
-        _buildRunner = buildRunner,
+        _nativeAssetsBuilder = nativeAssetsBuilder,
         _nativeAssetsYamlFile = nativeAssetsYamlFile,
         _analytics = analytics,
         super(
@@ -139,7 +138,7 @@ class HotRunner extends ResidentRunner {
   String? _sdkName;
   bool? _emulator;
 
-  NativeAssetsBuildRunner? _buildRunner;
+  final HotRunnerNativeAssetsBuilder? _nativeAssetsBuilder;
   final String? _nativeAssetsYamlFile;
 
   String? flavor;
@@ -374,17 +373,12 @@ class HotRunner extends ResidentRunner {
       nativeAssetsYaml = globals.fs.path.toUri(_nativeAssetsYamlFile);
     } else {
       final Uri projectUri = Uri.directory(projectRootPath);
-      _buildRunner ??= NativeAssetsBuildRunnerImpl(
-        projectUri,
-        debuggingOptions.buildInfo.packageConfig,
-        fileSystem,
-        globals.logger,
-      );
-      nativeAssetsYaml = await dryRunNativeAssets(
+      nativeAssetsYaml = await _nativeAssetsBuilder?.dryRun(
         projectUri: projectUri,
         fileSystem: fileSystem,
-        buildRunner: _buildRunner!,
         flutterDevices: flutterDevices,
+        logger: logger,
+        packageConfig: debuggingOptions.buildInfo.packageConfig,
       );
     }
 
@@ -1727,4 +1721,16 @@ class ReasonForCancelling {
   String toString() {
     return '$message.\nTry performing a hot restart instead.';
   }
+}
+
+/// An interface to enable overriding native assets build logic in other
+/// build systems.
+abstract class HotRunnerNativeAssetsBuilder {
+  Future<Uri?> dryRun({
+    required Uri projectUri,
+    required FileSystem fileSystem,
+    required List<FlutterDevice> flutterDevices,
+    required PackageConfig packageConfig,
+    required Logger logger,
+  });
 }
