+++ packages/flutter_tools/test/general.shard/isolated/build_system/targets/native_assets_test.dart
@@ -12,14 +12,14 @@ import 'package:flutter_tools/src/build_system/build_system.dart';
 import 'package:flutter_tools/src/build_system/exceptions.dart';
 import 'package:flutter_tools/src/build_system/targets/native_assets.dart';
 import 'package:flutter_tools/src/features.dart';
-import 'package:flutter_tools/src/native_assets.dart';
+import 'package:flutter_tools/src/isolated/native_assets/native_assets.dart';
 import 'package:native_assets_cli/native_assets_cli_internal.dart'
     as native_assets_cli;
 import 'package:package_config/package_config.dart' show Package;
 
-import '../../../src/common.dart';
-import '../../../src/context.dart';
-import '../../../src/fakes.dart';
+import '../../../../src/common.dart';
+import '../../../../src/context.dart';
+import '../../../../src/fakes.dart';
 import '../../fake_native_assets_build_runner.dart';
 
 void main() {
